Update fig to build in gcc, make compile under both c89 and c99 with -pedantic -Wall -Werror. Add Makefile. Kill fig_offset_t in favor of just using ptrdiff_t. Remember to check for preprocessor symbol existence before comparing them for compiler version checks. Fix dumb mistake where modern MSVC would use use vendor-specific types instead of stdint when it doesn't need to. Fix some places that were using printf but not casting the result beforehand.

+++ tests/fig_gif2ppm.c
@@ -45,12 +45,12 @@ int main(int argc, char **argv) {
             image_size = width * height;
             data = fig_image_get_canvas_data(image);
 
-            sprintf(buffer, "out.%03d.ppm", i);
+            sprintf(buffer, "out.%03d.ppm", (int) i);
 
             f = fopen(buffer, "wb");
 
             if(f != NULL) {
-                fprintf(f, "P6 %d %d 255 ", width, height);
+                fprintf(f, "P6 %d %d 255 ", (int) width, (int) height);
                 for(j = 0; j < image_size; ++j) {
                     fig_uint8_t out[3];
